refactor: 컴포넌트 경로 수정 및 새로운 레이아웃 컴포넌트 추가
