+++ mdx-components.tsx
@@ -4,6 +4,6 @@ import Image, { type ImageProps } from "next/image";
 export function useMDXComponents(components: MDXComponents) {
   return {
     ...components,
-    Image: (props: ImageProps) => <Image {...props} />,
+    Image: (props: ImageProps) => <Image {...props} alt="" />,
   };
 }
