+++ src/shared/components/layout/EmptyContent.tsx
