+++ src/shared/components/layout/Footer.tsx
