+++ src/shared/components/layout/SimpleLayout.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Container from "./Container";
+import Container from "@/shared/components/layout/Container";
 
 const SimpleLayout = ({
   title,
